DiscordGloom/requirements.txt
asyncio==3.4.3
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
idna==3.2
plyer==2.0.0
pyautogui==0.9.53
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
opencv-python==4.5.3.56
webbrowser==0.5.1
psutil
keyboard
pyperclip
pynput
pyinstaller
auto-py-to-exe
